fix broken test
Judging from the commit where the test was added the purpose was to test
that the Issuer constructor would accept absent kwargs, and the test was
probably written before it did so.

This removes the objective_disclosure kwarg which brings the number of
message parts down to 6

"_sd" only gets added to the payload when the its path contains ".", and
that only happens when object disclosures are added.  That could just be
a bug, but I'm inverting the test to assure the payload doesn't contain
"_sd" for now.